Refactor algorithms into its own module
